requirements.txt
asgiref==3.5.0
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
distro==1.6.0
Django==4.0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.3
jsonschema==3.2.0
paramiko==2.9.2
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dotenv==0.19.2
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
urllib3==1.26.8
websocket-client==0.59.0
django-environ==0.4.5
pre-commit==2.11.1
django-cors-headers
mysqlclient==2.0.3
uWSGI==2.0.19.1
docker-compose==1.29.2
djangorestframework==3.12.2
django-rest-knox==4.1.0
django-filter
factory-boy
coverage
codecov
attrs==21.4.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.11
codecov==2.1.12
coverage==6.3.1
cryptography==36.0.1
distlib==0.3.4
distro==1.6.0
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.4.5
django-filter==21.1
django-model-utils==4.2.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
factory-boy==3.2.1
Faker==12.1.0
filelock==3.4.2
identify==2.4.8
idna==3.3
jsonschema==3.2.0
mysqlclient==2.0.3
nodeenv==1.6.0
paramiko==2.9.2
platformdirs==2.4.1
pre-commit==2.11.1
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
request-id-django-log==0.2.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
toml==0.10.2
urllib3==1.26.8
uWSGI==2.0.19.1
virtualenv==20.13.0
websocket-client==0.59.0
django-import-export
drf-yasg
django-tables2